Convert ScrollLock to TS
