+++ src/components/ScrollLock/index.ts
@@ -1,0 +1,3 @@
+import ScrollLock from './ScrollLock'
+
+export default ScrollLock
